+++ lib/about.dart
@@ -1,15 +1,15 @@
 import 'dart:io';
 
-const String _about = """Goto v0.0.5
+const String _about = """Goto v0.0.6
 
 Copyright (c) 2020, Syed Mushaheed. All rights reserved.
 Licensed with BSD 3-Clause License.
 
 Written by Mushaheed Syed <smushaheed@gmail.com>""";
 
-/// Callback to show version information if -V or --version flag is used
+/// Callback to show version-about information if -V or --version flag is used
 void version(bool called) {
-  if (!called) return;
+  if (!called) return; // To prevent unexpected callbacks
   print(_about);
   exit(0);
 }