+++ lib/cmd.dart
@@ -4,6 +4,35 @@ import 'package:args/command_runner.dart';
 import 'package:goto/goto.dart';
 import 'package:goto/nerror.dart';
 
+const List<String> _illegalKeyNames = <String>[
+  'g',
+  'get',
+  'ls',
+  'l',
+  'list',
+  're',
+  'rename',
+  'rm',
+  'r',
+  'remove',
+  'save',
+  's',
+  'set',
+];
+
+void isKeyValid(String key) {
+  if (!RegExp(r'^\w*$').hasMatch(key)) {
+    GotoError.warn(
+        "'$key' has invalid format. Key must only contain alphabet, underscore, or numbers");
+    return;
+  }
+  if (_illegalKeyNames.contains(key)) {
+    GotoError.warn(
+        "'$key' is invalid. Key must not be the name or alias of any goto command.");
+    return;
+  }
+}
+
 /// When no other commands are executed then the argument specifies a &lt;key&gt;
 ///
 /// That &lt;key&gt; represents a path and if it's valid and exists, this will
@@ -24,12 +53,70 @@ class GoCommand {
   void run() {
     if (!valid) throw Exception('Invalid Arguments\n${parser.usage}');
     final String key = arguments.first;
-    Goto().gotoPath(key);
+    Goto.gotoPath(key);
+  }
+}
+
+abstract class ExtendedCommand extends Command<String> {
+  String get listAlias =>
+      aliases.isEmpty ? '' : '\nOther alias: ${aliases.join(', ')}\n';
+
+  @override
+  String get description;
+
+  @override
+  String get name;
+
+  /// A single-line template for how to invoke this command (e.g. "pub getpackage").
+  String get usageInvocation;
+
+  @override
+  String get invocation => '$usageInvocation$listAlias';
+}
+
+class RenameCommand extends ExtendedCommand {
+  // The [name] and [description] properties must be defined by every
+  // subclass.
+  @override
+  final name = "rename";
+  @override
+  final description = "Renames a key.";
+
+  RenameCommand();
+
+  @override
+  String get usageInvocation =>
+      'goto rename <previous_key_name> <new_key_name>';
+
+  @override
+  List<String> get aliases => ['re'];
+
+  // [run] may also return a Future.
+  @override
+  Future<String> run() {
+    if ((argResults.rest?.isEmpty ?? true) ||
+        (argResults.rest?.length ?? 0) < 2) {
+      GotoError.missing(usage);
+    }
+    final String previousKeyName = argResults.rest[0];
+    final String newKeyName = argResults.rest[1];
+    // Check if previousKeyExists, else throw KeyNotFound
+    if (!Goto.containsKey(previousKeyName)) {
+      GotoError("Found no path with '$previousKeyName' key");
+      return null;
+    }
+    if (previousKeyName == newKeyName) {
+      GotoError.exit('Error: Old & new key names must not be same.');
+    }
+    // If exists, remove that key and replace it with new key-name else.
+    isKeyValid(newKeyName);
+    print("Renaming key '$previousKeyName' with '$newKeyName'.. ");
+    Goto.rename(previousKeyName, newKeyName);
   }
 }
 
 /// Saves a path with key
-class SetCommand extends Command<String> {
+class SetCommand extends ExtendedCommand {
   // The [name] and [description] properties must be defined by every
   // subclass.
   @override
@@ -40,12 +127,10 @@ class SetCommand extends Command<String> {
   SetCommand();
 
   @override
-  String get invocation => 'goto set <key> <path>';
+  String get usageInvocation => 'goto set <key> <path>';
 
   @override
   List<String> get aliases => ['save', 's'];
-
-  final RegExp keyRegx = RegExp(r'^\w*$');
 
   // [run] may also return a Future.
   @override
@@ -54,23 +139,20 @@ class SetCommand extends Command<String> {
       GotoError.missing(usage);
     }
     String key = argResults.rest[0];
-    if (!keyRegx.hasMatch(key)) {
-      GotoError(
-          "Key invalid. Key must only contain alphabet, underscore, or numbers");
-    }
+    isKeyValid(key);
     String value = (argResults.rest.length == 1)
         ? Directory.current.path
         : argResults.rest[1];
     if (value == '.') {
       value = Directory.current.path;
     }
-    print('Saving $value with key $key.. ');
-    Goto().setKey(key, value);
+    print('Saving "$value" with key "$key".. ');
+    Goto.setKey(key, value);
   }
 }
 
 /// Retrieves and prints a path for a key
-class GetCommand extends Command<String> {
+class GetCommand extends ExtendedCommand {
   // The [name] and [description] properties must be defined by every
   // subclass.
   @override
@@ -81,7 +163,7 @@ class GetCommand extends Command<String> {
   GetCommand();
 
   @override
-  String get invocation => 'goto get <key>';
+  String get usageInvocation => 'goto get <key>';
 
   @override
   List<String> get aliases => ['g'];
@@ -95,16 +177,16 @@ class GetCommand extends Command<String> {
       GotoError.missing(usage);
     }
     String key = argResults.rest[0];
-    String reply = Goto().getPath(key);
+    String reply = Goto.getPath(key);
     if (reply == null) {
-      GotoError('Found no path with "$key" key');
+      GotoError("Found no path with '$key' key");
     }
     print('$key -> ${reply}');
   }
 }
 
 /// Remove a saved key-path or remove all saved key-paths
-class RemoveCommand extends Command<String> {
+class RemoveCommand extends ExtendedCommand {
   // The [name] and [description] properties must be defined by every
   // subclass.
   @override
@@ -119,7 +201,7 @@ class RemoveCommand extends Command<String> {
   }
 
   @override
-  String get invocation => 'goto remove <key>';
+  String get usageInvocation => 'goto remove <key>';
 
   @override
   List<String> get aliases => ['rm', 'r'];
@@ -129,7 +211,7 @@ class RemoveCommand extends Command<String> {
   Future<String> run() {
     if (argResults['all']) {
       // remove all
-      Goto().removeAll();
+      Goto.removeAll();
       return null;
     }
     if (argResults.rest?.isEmpty ?? true) {
@@ -137,12 +219,12 @@ class RemoveCommand extends Command<String> {
     }
     // [argResults] is set before [run()] is called and contains the options
     // passed to this command.
-    Goto().removeKey(argResults.rest[0]);
+    Goto.removeKey(argResults.rest[0]);
   }
 }
 
 /// Lists all saved key-path pairs in human readable format
-class ListCommand extends Command<String> {
+class ListCommand extends ExtendedCommand {
   // The [name] and [description] properties must be defined by every
   // subclass.
   @override
@@ -156,7 +238,7 @@ class ListCommand extends Command<String> {
   bool get takesArguments => false;
 
   @override
-  String get invocation => 'goto list';
+  String get usageInvocation => 'goto list';
 
   @override
   List<String> get aliases => ['ls', 'l'];
@@ -164,7 +246,7 @@ class ListCommand extends Command<String> {
   // [run] may also return a Future.
   @override
   Future<String> run() {
-    Map data = Goto().data;
+    Map data = Goto.data;
     if (data.isEmpty) {
       print('No saved records');
       return null;